requirements.txt
streamlit
torch
torchvision
torchaudio
grad-cam
lime
scikit-image
opencv-python
matplotlib
numpy
pandas
Pillow
python-dotenv
supabase
huggingface-hub